# documentation/api.yaml
openapi: 3.0.3
info:
  title: Westco Tack API Course Project Documentation.
  version: 1.0.11
  license:
    name: Apache 2.0
    url: "http://apache.org/"
servers:
  - url: https://educations-apis.herokuapp.com/
    description: Https Hosting Heroku
  - url: http://localhost:5000
    description: Dev Server on Port
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /user/getCourseOne/{id}:
    get:
      summary: Get one course user
      tags:
        - Course
      description: Get one course comment
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Get Course One for course Id
          example: 62b241cec57596d4147cf96c
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id: 
                    type: string
                  name: 
                    type: string
                  courseId:
                    type: string
                  userId: 
                    type: string
                  description:
                    type: string
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
              example:
                "_id": "62ae9ca688144d2d246e879d"
                "courseId": "62ae95af49ef0d92e405b231"
                "name": "Sunna"
                "description": "Assalamu Aleykum menga bu React Next JS darslari juda maqul bo'ldi "
                "userId": "62ac7f3fa525215eeede092f"
                "createdAt": "2022-06-19T03:48:54.490Z"
                "updatedAt": "2022-06-19T03:48:54.490Z"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /getCourseComment/{id}:
    get:
      summary: Get one course comment
      tags:
        - Course
      description: Get one course comment
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Get Course One for course Id
          example: 62ae9ca688144d2d246e879d
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                  name:
                    type:  string
                  authorId:
                    type:  array
                  price:
                    type: number
                  courseImg:
                    type: string
                  courseUsersId:
                    type: array
                  reating:
                    type: number
                  description:
                    type:  string
                  lessons: 
                    type: array
                  books:
                    type: array
                  comments:
                    type: array
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
              example:
                _id: 62ae948249ef0d92e405b22b
                name: Node.js darslari 0 dan pro gacha
                authorId: []
                price: 1500000
                courseImg: 18d181a6-d74b-40b3-b886-c409d76a6d21-file-nodejs_corse.png
                courseUsersId: []
                reating: 0
                description: ""
                lessons: []
                books: []
                comments: []
                createdAt: "2022-06-19T03:14:10.029Z"
                updatedAt: "2022-06-19T05:55:21.434Z"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /getCourseBook/{id}:
    get:
      summary: Get One Course Comment
      tags:
        - Course
      description: Get one Course Comment
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Get Course One for course Id
          example: 62ae9ca688144d2d246e879d
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id: string
                  name: string
                  courseId: string
                  books: string
                  urlBooks: string
                  description: string
                  createdAt: string
                  updatedAt: string
              example:
                _id: string
                name: string
                courseId: string
                books: string
                urlBooks: string
                description: string
                createdAt: string
                updatedAt: string
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
...
  /user/cuourseEnrole/{id}:
    put:
      summary: User Course enrole
      tags:
        - User Course
      description: We add the user to the course and we add the user ID to the course
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: this is the course ID 
          example: 62ae948249ef0d92e405b22b
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                userId: 
                  type: number
                userPay: 
                  type: number
            example:
              userId: "62af2d6316c2ba27f61eb8df"
              userPay: "1500000"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  userEnroled:
                    type: string
                example:
                  userEnroled: "User sucsess fuly enrolled cours"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /user/getCourseProfile/{id}:
    get:
      summary: Get One User
      tags:
        - User Course
      description: Get user courses
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Get Courses for user Id
          example: 62b23b7a7b947f6c0c88245b
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id: string
                  name: string
                  authorId: array
                  price: number
                  courseImg: string
                  courseUsersId: array
                  reating: number
                  description: string
                  lessons: array
                  books: array
                  comments: array
                  createdAt: string
                  updatedAt: string
              example:
                _id: 62ae948249ef0d92e405b22b
                name: Node.js darslari 0 dan pro gacha
                authorId: []
                price: 1500000
                courseImg: 18d181a6-d74b-40b3-b886-c409d76a6d21-file-nodejs_corse.png
                courseUsersId: []
                reating: 0
                description: ""
                lessons: []
                books: []
                comments: []
                createdAt: "2022-06-19T03:14:10.029Z"
                updatedAt: "2022-06-19T05:55:21.434Z"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /user/getCourseOneProfile/{id}:
    get:
      summary: User Profile get course One
      tags:
        - User Course
      description: Get one Course
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Get Course One for course Id
          example: 62aec4b4edf4613e87f85cb6
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id: string
                  name: string
                  authorId: array
                  price: number
                  courseImg: string
                  courseUsersId: array
                  reating: number
                  description: string
                  lessons: array
                  books: array
                  comments: array
                  createdAt: string
                  updatedAt: string
              example:
                _id: 62ae948249ef0d92e405b22b
                name: Node.js darslari 0 dan pro gacha
                authorId: []
                price: 1500000
                courseImg: 18d181a6-d74b-40b3-b886-c409d76a6d21-file-nodejs_corse.png
                courseUsersId: []
                reating: 0
                description: ""
                lessons: []
                books: []
                comments: []
                createdAt: "2022-06-19T03:14:10.029Z"
                updatedAt: "2022-06-19T05:55:21.434Z"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error

  # /user/courseRate/{id}:
  #   post:
  #     summary: User Profile get course One
  #     tags:
  #       - User Course
  #     description: Get one Course
  #     parameters:
  #       - in: path
  #         name: id
  #         schema:
  #           type: string
  #         required: true
  #         description: course id for the user to rate
  #         example: 62aec4b4edf4613e87f85cb6
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             type: object
  #             properties:
  #               userId:
  #                 type: string
  #               rating:
  #                 type: string
  #             example:
  #               userId: 62b216049920204eb56c309b
  #               rating: {5:1}
  #     responses:
  #       200:
  #         description: OK
  #         content:
  #           application/json:
  #             schema:
  #               type: object
  #             example:
  #               msg: "kurga baho qo'ydingiz"
  #       404:
  #         description: Not Found Course
  #       500:
  #         description: Internal Server Error
  /user/userCourseDelete/{id}:
    put:
      summary: User course delete
      tags:
        - User Course
      description: We add the user to the course and we add the user ID to the course
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: this is the course ID to add to user cursors
          example: 62ae948249ef0d92e405b22b
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
              example:
                userId: "62af2d6316c2ba27f61eb8df"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  userEnroled:
                    type: string
              example:
                userEnroled: "User sucsess fuly enrolled cours"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /user/profileUpdate:
    put:
      summary: Super Admin Update Profile
      tags:
        - User Course
      description: To change the Super Admin cabinet data, you must first enter the old password and the security word and then the new ones
      consumes:
        - multipart/form-data
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                userId:
                  type: string
                name:
                  type: string
                email:
                  type: string
                oldPassword:
                  type: string
                newPassword:
                  type: string
                userAvatar:
                  type: string
                  format: binary
      responses:
        200:
          description: OK
          content:
            multipart/form-data:
              schema:
                type: object
                properties:
                  adminProfile: string
              example:
                adminProfile: Secsess fully updated User Profile
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /user/forgotPassword/searchUser:
    post:
      summary: User forget password update
      tags:
        - User Course
      description: When you forget the password, we confirm that the user is registered by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
              example:
                email: Sunna@gmail.com
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: integer
                  _id: integer
                  name: string
                  email: string
                  password: string
                  secretInfo: string
                  createdAt: string
                  updatedAt: string
              example:
                _id: 62af3d203816e616bd006514
                name: Sunnatullox
                email: Sunnatullox@gmail.com
                password: $2a$10$vF7DFPtLrXo4FXpO2vm7U.GXPHpmJWgVWD51rVsXVgL84cAzjyPrW
                secretInfo: $2a$10$p/oRKgRopkRvx5/a/MTp/OEnhVk7wx9i8SIKJkVFPPipQGUUBSX2S
                createdAt: 2022-06-19T15:13:36.493+00:00
                updatedAt: 2022-06-20T14:01:43.788+00:00
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /user/fotgetPassword/updatedPasword:
    post:
      summary: User forget password update
      tags:
        - User Course
      description: When you forget the password, we confirm that the user is registered by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                password:
                  type: string
              example:
                userId: 62af2d6316c2ba27f61eb8df
                password: "12345"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id: integer
                  name: string
                  email: string
                  password: string
                  secretInfo: string
                  createdAt: string
                  updatedAt: string
              example:
                _id: 62af3d203816e616bd006514
                name: Sunnatullox
                email: Sunnatullox@gmail.com
                password: $2a$10$vF7DFPtLrXo4FXpO2vm7U.GXPHpmJWgVWD51rVsXVgL84cAzjyPrW
                secretInfo: $2a$10$p/oRKgRopkRvx5/a/MTp/OEnhVk7wx9i8SIKJkVFPPipQGUUBSX2S
                createdAt: 2022-06-19T15:13:36.493+00:00
                updatedAt: 2022-06-20T14:01:43.788+00:00
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
...